Rest_api/requirements.txt
asgiref==3.5.1
asttokens==2.0.5
backcall==0.2.0
colorama==0.4.4
decorator==5.1.1
Django==3.2.12
django-extensions==3.1.5
djangorestframework==3.13.1
executing==0.8.3
ipython==8.3.0
jedi==0.18.1
matplotlib-inline==0.1.3
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.29
pure-eval==0.2.2
Pygments==2.12.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.2.0
wcwidth==0.2.5
dj-rest-auth==2.2.4
django-allauth==0.50.0
django-cors-headers==3.12.0